+++ src/Details/Details.tsx
@@ -1,0 +1,10 @@
+
+function Details() {
+  return (
+    <div>
+      <h1>Details</h1>
+    </div>
+  );
+}
+
+export default Details;
