+++ src/Details/index.tsx
@@ -1,10 +1,3 @@
-
-function Details() {
-  return (
-    <div>
-      <h1>Details</h1>
-    </div>
-  );
-}
+import Details from "./Details";
 
 export default Details;
