+++ src/Home/Home.tsx
@@ -1,0 +1,10 @@
+
+function Home() {
+  return (
+    <div>
+      <h1>Home</h1>
+    </div>
+  );
+}
+
+export default Home;
