+++ src/Home/index.tsx
@@ -1,10 +1,3 @@
-
-function Home() {
-  return (
-    <div>
-      <h1>Home</h1>
-    </div>
-  );
-}
+import Home from "./Home";
 
 export default Home;
