+++ src/Login/Login.tsx
@@ -1,0 +1,10 @@
+
+function Login() {
+  return (
+    <div>
+      <h1>Login</h1>
+    </div>
+  );
+}
+
+export default Login;
