+++ src/Profile/index.tsx
@@ -1,21 +1,3 @@
-import { useParams } from "react-router";
-
-function Profile() {
-  const { uid } = useParams();
-
-  if (!uid) {
-    return (
-      <div>
-        <h1>Current User's Profile</h1>
-      </div>
-    );
-  } else {
-    return (
-      <div>
-        <h1>User {uid}'s Profile</h1>
-      </div>
-    );
-  }
-}
+import Profile from "./Profile";
 
 export default Profile;
