+++ src/Search/Search.tsx
@@ -1,0 +1,10 @@
+
+function Search() {
+  return (
+    <div>
+      <h1>Search</h1>
+    </div>
+  );
+}
+
+export default Search;
